make array `move` a util
